Added automated tests, just run test.sh in aut-test/ directory
each test $TEST is defined by three files

$TEST_input.txt
$TEST_output.txt
$TEST_errors.txt

the bash script test.sh runs each *input.txt file and compares (using
the diff command) their stdout and sterr with the given *output.txt
and *errors.txt. If they match the test is succesful, if not it fails
and shows the diff of that output with the expected output.

change or add files in the auto-test/usecase/ to change or add file
new tests. Right now they tests executed are hardcode in test.sh so
you have to add the names of the new test there if you want extra
tests.